// frontend/src/lib/api/model/errorDetail.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Orvo API
 * OpenAPI spec version: 1.0.0
 */

export interface ErrorDetail {
  /** Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id' */
  location?: string;
  /** Error message text */
  message?: string;
  /** The value at the given location */
  value?: unknown;
}