// frontend/src/lib/api/model/serviceRemediationMapping.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Orvo API
 * OpenAPI spec version: 1.0.0
 */

export interface ServiceRemediationMapping {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  created_at: string;
  created_by_user_id: string;
  id: string;
  organization_id: string;
  repository_full_name: string;
  repository_id: string;
  service_name: string;
  updated_at: string;
  updated_by_user_id: string;
}